fix bounding box function, get rid of extensions, copied from imported libraries
